Move all data structures/code in lib.rs to common.rs to be in sync with agave.
