+++ votor/src/consensus_pool/slot_stake_counters.rs
@@ -1,8 +1,12 @@
 use {
     crate::{
-        consensus_pool::stats::ConsensusPoolStats, event::VotorEvent, Stake,
-        SAFE_TO_NOTAR_MIN_NOTARIZE_AND_SKIP, SAFE_TO_NOTAR_MIN_NOTARIZE_FOR_NOTARIZE_OR_SKIP,
-        SAFE_TO_NOTAR_MIN_NOTARIZE_ONLY, SAFE_TO_SKIP_THRESHOLD,
+        common::{
+            Stake, SAFE_TO_NOTAR_MIN_NOTARIZE_AND_SKIP,
+            SAFE_TO_NOTAR_MIN_NOTARIZE_FOR_NOTARIZE_OR_SKIP, SAFE_TO_NOTAR_MIN_NOTARIZE_ONLY,
+            SAFE_TO_SKIP_THRESHOLD,
+        },
+        consensus_pool::stats::ConsensusPoolStats,
+        event::VotorEvent,
     },
     solana_hash::Hash,
     solana_votor_messages::vote::Vote,
