+++ votor/src/consensus_pool_service.rs
@@ -6,13 +6,13 @@ mod stats;
 use {
     crate::{
         commitment::{update_commitment_cache, CommitmentAggregationData, CommitmentType},
+        common::DELTA_STANDSTILL,
         consensus_pool::{
             parent_ready_tracker::BlockProductionParent, AddVoteError, ConsensusPool,
         },
         event::{LeaderWindowInfo, VotorEvent, VotorEventSender},
         voting_service::BLSOp,
         votor::Votor,
-        DELTA_STANDSTILL,
     },
     crossbeam_channel::{select, Receiver, Sender, TrySendError},
     solana_clock::Slot,
